+++ docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 * You can now add Fashion (shaders and ornaments) to loadouts. Creating a loadout from equipped automatically adds in the current shaders and ornaments, and you can edit them manually from the Loadouts editor.
 * Consumables can now be pulled from postmasters other than the active character's.
+* Vendors page now correctly recognizes owned bounties per character and is more accurate about mods ownership.
+* Fixed an issue that could make moving searches containing stacks of items to fail.
+* Fixes for display on iPhones with rounded corners and a notch.
 
 ## 6.97.3 <span class="changelog-date">(2021-12-30)</span>